Change plotting package from 'plot' to 'plots'
